requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
PyYAML==6.0
requests==2.27.1
tqdm==4.64.0
urllib3==1.26.9
